% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_pres}
\alias{convert_pres}
\title{Convert between common pressure units}
\usage{
convert_pres(x, from, to)
}
\arguments{
\item{x}{A numeric vector of pressures to convert.}

\item{from, to}{The units to convert from and to. Must be one of \code{"hpa"},
\code{"mb"}, \code{"inhg"}, or \code{"mmhg"}. These options are case-insensitive.}
}
\value{
A numeric vector of the same length as \code{x} containing the
converted values. \code{NA} values are preserved.
}
\description{
This function converts between common pressure units: hectopascals (hPa),
millibars (mb), inches of mercury (inHg), or millimeters of mercury (mmHg).
Note that hectopascals and millibars are equivalent units, so no conversion
is needed between them.
}
\examples{
# Convert 1013.25 hPa to inches of mercury
convert_pres(1013.25, "hpa", "inhg")
}
\seealso{
Other meteorological unit conversion functions: 
\code{\link{convert_prcp}()},
\code{\link{convert_temp}()},
\code{\link{convert_wd}()},
\code{\link{convert_ws}()}
}
\concept{meteorological unit conversion functions}
