% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_temp}
\alias{convert_temp}
\title{Convert between common temperature units}
\usage{
convert_temp(x, from, to)
}
\arguments{
\item{x}{A numeric vector of temperatures to convert.}

\item{from, to}{The units to convert from and to. Must be one of \code{"c"}
(Celsius), \code{"f"} (Fahrenheit), or \code{"k"} (Kelvin). These options are
case-insensitive.}
}
\value{
A numeric vector of the same length as \code{x} containing the
converted values. \code{NA} values are preserved.
}
\description{
This function converts between Celsius, Fahrenheit, and Kelvin temperature
units.
}
\examples{
# Convert 100 degrees Celsius to Fahrenheit
convert_temp(100, "c", "f")
}
\seealso{
Other meteorological unit conversion functions: 
\code{\link{convert_prcp}()},
\code{\link{convert_pres}()},
\code{\link{convert_wd}()},
\code{\link{convert_ws}()}
}
\concept{meteorological unit conversion functions}
