% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_wd}
\alias{convert_wd}
\title{Convert between common wind direction units}
\usage{
convert_wd(x, from, to, resolution = 8)
}
\arguments{
\item{x}{A numeric vector of wind directions to convert. If \code{from} is
"compass", this should be a character vector of compass directions (e.g.,
"N", "NE", etc.).}

\item{from, to}{The units to convert from and to. Must be one of \code{"degrees"},
\code{"radians"}, or \code{"compass"}. These options are case-insensitive.}

\item{resolution}{The number of compass points to use when converting to or
from compass directions. Must be one of \code{4}, \code{8}, \code{16}, or \code{32}. The
default is \code{8}, which corresponds to the standard 8-point compass
(N/NE/E/SE/S/SW/W/NW).}
}
\value{
A numeric vector of the same length as \code{x} containing the converted
values, or a factor vector if \code{to = "compass"}. \code{NA} values are preserved.
}
\description{
This function converts between common wind direction units: degrees and
compass directions. The \code{resolution} argument specifies the number of compass
points to use when converting to or from compass directions (e.g., \code{4} for
N/E/S/W, \code{8} for N/NE/E/SE/S/SW/W/NW, etc.). Note that conversions to and
from compass directions are 'destructive' in that they involve binning wind
direction, which reduces precision.
}
\examples{
# Convert 90 degrees to compass direction with 8-point resolution
convert_wd(90, "degrees", "compass")
}
\seealso{
Other meteorological unit conversion functions: 
\code{\link{convert_prcp}()},
\code{\link{convert_pres}()},
\code{\link{convert_temp}()},
\code{\link{convert_ws}()}
}
\concept{meteorological unit conversion functions}
