% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_ws}
\alias{convert_ws}
\title{Convert between common wind speed units}
\usage{
convert_ws(x, from, to)
}
\arguments{
\item{x}{A numeric vector of wind speeds to convert.}

\item{from, to}{The units to convert from and to. Must be one of \code{"m/s"},
\code{"km/h"}, \code{"mph"}, \code{"knots"}, \code{"ft/s"}, or \code{"beaufort"}. These options are
case-insensitive.}
}
\value{
A numeric vector of the same length as \code{x} containing the
converted values. \code{NA} values are preserved.
}
\description{
This function converts between common wind speed units: meters per second
(m/s), kilometers per hour (km/h), miles per hour (mph), knots, feet per
second (ft/s), or beaufort. Note that conversions to and from the Beaufort
scale are 'destructive' in that they involve binning wind speed, which
reduces precision.
}
\examples{
# Convert 10 meters per second to kilometers per hour
convert_ws(10, "m/s", "km/h")
}
\seealso{
Other meteorological unit conversion functions: 
\code{\link{convert_prcp}()},
\code{\link{convert_pres}()},
\code{\link{convert_temp}()},
\code{\link{convert_wd}()}
}
\concept{meteorological unit conversion functions}
