% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{exportADMS}
\alias{exportADMS}
\title{Deprecated data functions}
\usage{
exportADMS(dat, out = "./ADMS_met.MET", interp = FALSE, maxgap = 2)
}
\arguments{
\item{dat}{A data frame imported by \code{\link[=importNOAA]{importNOAA()}}.}

\item{out}{A file name for the ADMS file. The file is written to the working
directory by default.}

\item{interp}{Should interpolation of missing values be undertaken? If \code{TRUE}
linear interpolation is carried out for gaps of up to and including
\code{maxgap}.}

\item{maxgap}{The maximum gap in hours that should be interpolated where
there are missing data when \code{interp = TRUE.} Data with gaps more than
\code{maxgap} are left as missing.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is part of an old \code{worldmet} API. Please use the following
alternatives:
\itemize{
\item \code{\link[=exportADMS]{exportADMS()}} -> \code{\link[=write_adms]{write_adms()}}
\item the \code{path} argument -> \code{\link[=write_met]{write_met()}}
}
}
