% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Find a ISD site code and other meta data}
\usage{
getMeta(site = "heathrow", lat = NA, lon = NA, n = 10)
}
\arguments{
\item{site}{A site name search string e.g. \code{site =
"heathrow"}. The search strings and be partial and can be upper or
lower case e.g. \code{site = "HEATHR"}.}

\item{lat}{A latitude in decimal degrees to search. Takes the
values -90 to 90.}

\item{lon}{A longitude in decimal degrees to search. Takes values
-180 to 180. Negative numbers are west of the Greenwich meridian.}

\item{n}{The number of nearest sites to search based on
\code{latitude} and \code{longitude}.}
}
\value{
A data frame is returned with all available meta data,
mostly importantly including a \code{code} that can be supplied to
\code{\link{importNOAA}}. If latitude and longitude searches are
made an approximate distance, \code{dist} in km is also returned.
}
\description{
Get information on meteorological sites
}
\details{
This function is primarily used to find a site code that can be
used to access data using \code{\link{importNOAA}}. Sites searches
of approximately 30,000 sites can be carried out based on the site
name and based on the nearest locations based on user-supplied
latitude and logitude.
}
\examples{
\dontrun{
## search for sites with name beijing
getMeta(site = "beijing")
}

\dontrun{
## search for near a specified lat/lon - near Beijing airport
## returns 'n' nearest by default
getMeta(lat = 40, lon = 116.9)
}
}
\author{
David Carslaw
}

