% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMetaLive}
\alias{getMetaLive}
\title{Obtain site meta data from NOAA server}
\usage{
getMetaLive(...)
}
\arguments{
\item{...}{Currently unused.}
}
\value{
A tibble with meta data.
}
\description{
Obtain site meta data from NOAA server
}
\examples{
\dontrun{
meta <- getMetaLive()
head(meta)
}
}
