% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ghcnh_meta.R
\name{import_ghcnh_codes}
\alias{import_ghcnh_codes}
\title{Import Country and State Codes used in the Global Historical Climatology
Network hourly (GHCNh) Database}
\usage{
import_ghcnh_codes(table = c("countries", "states"))
}
\arguments{
\item{table}{One of \code{"countries"} or \code{"states"}, specifying whether country
or state codes wish to be returned by the function.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Import Country and State Codes used in the Global Historical Climatology
Network hourly (GHCNh) Database
}
\author{
Jack Davison
}
