% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGHCNdata.R
\name{importGHCNdaily}
\alias{importGHCNdaily}
\title{Import Global Historical Climatology Network daily (GHCNd) Data}
\usage{
importGHCNdaily(code, year, all = FALSE)
}
\arguments{
\item{code}{The specific station code(s) of interest, identified using
\code{\link[=importGHCNstations]{importGHCNstations()}}.}

\item{year}{The specific year(s) to import.}

\item{all}{GHCN data comes bundled with various extra columns, such as
measurement codes, quality codes, report types, source codes, attributes,
and so on. By default, these are removed. Setting \code{all = TRUE} will retain
these in the returned table.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function imports daily data from the GHCNd.
}
\author{
Jack Davison
}
