% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGHCNmeta.R
\name{importGHCNstations}
\alias{importGHCNstations}
\title{Import Global Historical Climatology Network (GHCN) Metadata}
\usage{
importGHCNstations(
  source = c("hourly", "daily"),
  lat = NULL,
  lng = NULL,
  n = 10,
  crs = 4326,
  station = NULL,
  country = NULL,
  state = NULL,
  return = c("map", "table", "sf")
)
}
\arguments{
\item{source}{The GHCN contains several different datasets, including
"hourly" and "daily" measurements. \code{source} will return the metadata for
sites included in each of these data sources, defaulting to \code{"hourly"}.}

\item{lat, lng}{Values representing a decimal latitude and longitude (or other
Y/X coordinate if using a different \code{crs}). If provided,
\code{importGHCNstations()} will locate \code{n} stations near to this global
coordinate.}

\item{n}{The number of nearest sites to search based on \code{lat} and \code{lng}.
Defaults to \code{10}.}

\item{crs}{The coordinate reference system (CRS) of \code{lat} and \code{lng}, passed
to \code{sf::st_crs()}. By default this is \href{https://epsg.io/4326}{EPSG:4326},
the CRS associated with the commonly used latitude and longitude
coordinates. Different coordinate systems can be specified using crs (e.g.,
\code{crs = 27700} for the \href{https://epsg.io/27700}{British National Grid}).}

\item{station}{A string which which to search and filter station names. For
example, \code{station = "heathrow"} would return all stations containing the
word "heathrow". The strings can be partial and can be upper or lower case
e.g. \code{site = "HEATHR"}.}

\item{country, state}{One or more two-letter 'country' or 'state' codes with
which to filter the metadata. These can be obtained with
\code{\link[=importGHCNcodes]{importGHCNcodes()}}.}

\item{return}{The form in which to return the data. One of:
\itemize{
\item \code{"map"} (the default), to return an interactive \code{leaflet} map.
\item \code{"table"}, to return a \link[tibble:tibble-package]{tibble}
\item \code{"sf"}, to return an \link[sf:sf]{sf spatial table}
}}
}
\value{
Varies (see the \code{return} parameter).
}
\description{
This function imports a table of metadata outlining the available sites in
the GHCN.
}
\author{
Jack Davison
}
