% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metNOAA.R
\name{importNOAA}
\alias{importNOAA}
\title{Import meteorological data}
\usage{
importNOAA(code = "037720-99999", year = 2014)
}
\arguments{
\item{code}{The identifing code as a character string. The code is
a combination of the USAF and the WBAN unique identifiers. The
codes are sperated by a \dQuote{-} e.g. \code{code =
"037720-99999"}.}

\item{year}{The year to import. This can be a vector of years
e.g. \code{year = 2000:2005}.}
}
\value{
Returns a data frame of surface observations. The data
frame is consistent for use with the \code{openair} package.
}
\description{
Main function for importing meteorological data
}
\details{
This is the main function to import data from the NOAA Integrated
Surface Database (ISD). The ISD contains detailed surface
meteorological data from around the world for over 30,000
locations. For general information of the ISD see
\url{https://www.ncdc.noaa.gov/isd} and the map here
\url{https://gis.ncdc.noaa.gov/map/viewer/#app=cdo&cfg=cdo&theme=hourly&layers=1}.
}
\examples{
\dontrun{
## use Beijing airport code (see getMeta example)
dat <- importNOAA(code = "545110-99999", year = 2010:2011)
}
}
\author{
David Carslaw
}
\seealso{
\code{\link{getMeta}} to obtain the codes based on
various site search approaches.
}

