% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_adms.R
\name{read_adms}
\alias{read_adms}
\title{Read a file in ADMS format}
\usage{
read_adms(file, progress = rlang::is_interactive())
}
\arguments{
\item{file}{One or more file paths to ADMS \code{.MET} files. If multiple files
are specified, they will be bound together into a single \code{data.frame}.}

\item{progress}{Show a progress bar when reading many files? Defaults to
\code{TRUE} in interactive R sessions. Passed to \code{.progress} in \code{\link[purrr:map]{purrr::map()}}.}
}
\description{
This function reads a file in ADMS format (e.g., a file written by
\code{\link[=write_adms]{write_adms()}}) as an R \code{data.frame}. This could be useful to check or
analyse ADMS files before they are used.
}
\seealso{
Other ADMS functions: 
\code{\link{write_adms}()}
}
\author{
Jack Davison
}
\concept{ADMS functions}
