% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldmet-package.R
\docType{package}
\name{worldmet}
\alias{worldmet}
\alias{worldmet-package}
\title{Access surface meteorological data from the NOAA Integrated Surface Database
from around the world}
\description{
This package contains functions to import surface meteorolgical data from
 over 30,000 sites around the world. These data are curated by NOAA as part
 of the Integrated Surface Database.
}
\author{
David Carslaw
}
\references{
For general information of the ISD see \url{https://www.ncdc.noaa.gov/isd} and the map here \url{https://gis.ncdc.noaa.gov/map/viewer/#app=cdo&cfg=cdo&theme=hourly&layers=1}.
}
\seealso{
See \url{https://github.com/davidcarslaw/openair} for information on the related \code{openair} package.
}
\keyword{methods}

