% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-isd.R
\name{getMeta}
\alias{getMeta}
\alias{getMetaLive}
\alias{importNOAA}
\title{Deprecated ISD import functions}
\usage{
getMeta(
  site = "heathrow",
  lat = NA,
  lon = NA,
  country = NA,
  state = NA,
  n = 10,
  end.year = "current",
  provider = c("OpenStreetMap", "Esri.WorldImagery"),
  plot = TRUE,
  returnMap = FALSE
)

getMetaLive(...)

importNOAA(
  code = "037720-99999",
  year = 2014,
  hourly = TRUE,
  n.cores = 1,
  quiet = FALSE,
  path = NA
)
}
\arguments{
\item{site}{A site name search string e.g. \code{site = "heathrow"}. The search
strings and be partial and can be upper or lower case e.g. \code{site = "HEATHR"}.}

\item{lat}{A latitude in decimal degrees to search. Takes the values -90 to
90.}

\item{lon}{A longitude in decimal degrees to search. Takes values -180 to
180. Negative numbers are west of the Greenwich meridian.}

\item{country}{The country code. This is a two letter code. For a full
listing see \url{https://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.csv}.}

\item{state}{The state code. This is a two letter code.}

\item{n}{The number of nearest sites to search based on \code{latitude} and
\code{longitude}.}

\item{end.year}{To help filter sites based on how recent the available data
are. \code{end.year} can be "current", "any" or a numeric year such as 2016, or
a range of years e.g. 1990:2016 (which would select any site that had an
end date in that range. \strong{By default only sites that have some data for the
current year are returned}.}

\item{provider}{By default a map will be created in which readers may toggle
between a vector base map and a satellite/aerial image. \code{provider} allows
users to override this default; see
\url{http://leaflet-extras.github.io/leaflet-providers/preview/} for a list
of all base maps that can be used. If multiple base maps are provided, they
can be toggled between using a "layer control" interface.}

\item{plot}{If \code{TRUE} will plot sites on an interactive leaflet map.}

\item{returnMap}{Should the leaflet map be returned instead of the meta data?
Default is \code{FALSE}.}

\item{...}{Currently unused.}

\item{code}{The identifying code as a character string. The code is a
combination of the USAF and the WBAN unique identifiers. The codes are
separated by a \dQuote{-} e.g. \code{code = "037720-99999"}.}

\item{year}{The year to import. This can be a vector of years e.g. \code{year = 2000:2005}.}

\item{hourly}{Should hourly means be calculated? The default is \code{TRUE}. If
\code{FALSE} then the raw data are returned.}

\item{n.cores}{Number of cores to use for parallel processing. Default is 1
and hence no parallelism.}

\item{quiet}{If FALSE, print missing sites / years to the screen.}

\item{path}{If a file path is provided, the data are saved as an rds file at
the chosen location e.g.  \code{path = "C:/Users/David"}. Files are saved by
year and site.}
}
\description{
Up until 2024, NOAA supported the Integrated Surface Database (ISD). This has
been superseded by the Global Historical Climatology Network hourly (GHCNh).
The following three functions relate to the ISD, but have been superseded.
\itemize{
\item \code{getMeta()} (obtains ISD metadata)
\item \code{getMetaLive()} (obtains the "raw" form of the ISD metadata)
\item \code{importNOAA()} (imports hourly meteorological data)
}
}
\author{
David Carslaw
}
