% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportADMS.R
\name{exportADMS}
\alias{exportADMS}
\title{Export a meteorological data frame in ADMS format}
\usage{
exportADMS(dat, out = "./ADMS_met.MET")
}
\arguments{
\item{dat}{A data frame imported by \code{\link{importNOAA}}.}

\item{out}{A file name for the ADMS file. The file is written in the working directory by default.}
}
\value{
Writes a text file to a location of the user's choosing.
}
\description{
Export a meteorological data frame in ADMS format
}

