% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ghcn_daily.R
\name{import_ghcn_daily}
\alias{import_ghcn_daily}
\title{Import data from the Global Historical Climatology daily (GHCNd) database}
\usage{
import_ghcn_daily(
  station,
  append_codes = FALSE,
  progress = rlang::is_interactive()
)
}
\arguments{
\item{station}{One or more site codes for the station(s) of interest,
obtained using \code{\link[=import_ghcn_stations]{import_ghcn_stations()}}.}

\item{append_codes}{Logical. Should attribute codes be appended to the output
dataframe?}

\item{progress}{Show a progress bar when importing many stations/years?
Defaults to \code{TRUE} in interactive R sessions. Passed to \code{.progress} in
\code{\link[purrr:map]{purrr::map()}} and/or \code{\link[purrr:pmap]{purrr::pmap()}}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function flexibly accesses meteorological data from the GHCNd database.
Users can provide any of stations, and control whether attribute codes are
returned with the data.
}
\seealso{
Other GHCN functions: 
\code{\link{import_ghcn_countries}()},
\code{\link{import_ghcn_hourly}()},
\code{\link{import_ghcn_inventory}()},
\code{\link{import_ghcn_monthly_temp}()},
\code{\link{import_ghcn_stations}()}
}
\author{
Jack Davison
}
\concept{GHCN functions}
