% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ghcn_daily.R
\name{import_ghcn_daily}
\alias{import_ghcn_daily}
\title{Import data from the Global Historical Climatology daily (GHCNd) database}
\usage{
import_ghcn_daily(
  station,
  append_codes = FALSE,
  progress = rlang::is_interactive()
)
}
\arguments{
\item{station}{One or more site codes for the station(s) of interest,
obtained using \code{\link[=import_ghcn_stations]{import_ghcn_stations()}}.}

\item{append_codes}{Logical. Should attribute codes be appended to the output
dataframe?}

\item{progress}{Show a progress bar when importing many stations/years?
Defaults to \code{TRUE} in interactive R sessions. Passed to \code{.progress} in
\code{\link[purrr:map]{purrr::map()}} and/or \code{\link[purrr:pmap]{purrr::pmap()}}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function flexibly accesses meteorological data from the GHCNd database.
Users can provide any of stations, and control whether attribute codes are
returned with the data.
}
\section{Parallel Processing}{


If you are importing a lot of meteorological data, this can take a long
while. This is because each combination of year and station requires
downloading a separate data file from NOAA's online data directory, and the
time each download takes can quickly add up. Many data import functions in
\code{{worldmet}} can use parallel processing to speed downloading up, powered
by the capable \code{{mirai}} package. If users have any \code{{mirai}} "daemons"
set, these functions will download files in parallel. The greatest benefits
will be seen if you spawn as many daemons as you have cores on your
machine, although one fewer than the available cores is often a good rule
of thumb. Your mileage may vary, however, and naturally spawning more
daemons than station-year combinations will lead to diminishing returns.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set workers - once per session
mirai::daemons(4)

# import lots of data - NB: no change in the import function!
big_met <- import_ghcn_hourly(code = "UKI0000EGLL", year = 2010:2025)
}\if{html}{\out{</div>}}
}

\seealso{
Other GHCN functions: 
\code{\link{import_ghcn_countries}()},
\code{\link{import_ghcn_hourly}()},
\code{\link{import_ghcn_inventory}()},
\code{\link{import_ghcn_monthly_temp}()},
\code{\link{import_ghcn_stations}()}
}
\author{
Jack Davison
}
\concept{GHCN functions}
