% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ghcn_data.R, R/import_ghcnh_data.R
\name{import_ghcn_hourly}
\alias{import_ghcn_hourly}
\title{Import Global Historical Climatology Network hourly (GHCNh) Data}
\usage{
import_ghcn_hourly(code, year, all = FALSE)

import_ghcn_hourly(code, year, all = FALSE)
}
\arguments{
\item{code}{The specific station code(s) of interest, identified using
\code{\link[=import_ghcn_stations]{import_ghcn_stations()}}.}

\item{year}{The specific year(s) to import.}

\item{all}{GHCNh data comes bundled with various extra columns, such as
measurement codes, quality codes, report types, and source codes. By
default, these are removed. Setting \code{all = TRUE} will retain these in the
returned table.}
}
\value{
a \link[tibble:tibble-package]{tibble}

a \link[tibble:tibble-package]{tibble}
}
\description{
This function imports hourly data from the GHCNh.

This function imports hourly data from the GHCNh.
}
\author{
Jack Davison
}
