% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ghcn_hourly.R
\name{import_ghcn_hourly}
\alias{import_ghcn_hourly}
\title{Import data from the Global Historical Climatology hourly (GHCNh) database}
\usage{
import_ghcn_hourly(
  station = "UKI0000EGLL",
  year = NULL,
  source = c("psv", "parquet"),
  hourly = TRUE,
  extra = FALSE,
  abbr_names = TRUE,
  append_codes = FALSE,
  codes = c("measurement_code", "quality_code", "report_type", "source_code",
    "source_id"),
  progress = rlang::is_interactive()
)
}
\arguments{
\item{station}{One or more site codes for the station(s) of interest,
obtained using \code{\link[=import_ghcn_stations]{import_ghcn_stations()}}.}

\item{year}{One or more years of interest. If \code{NULL}, the default, all years
of data available for the chosen \code{station}s will be imported.}

\item{source}{There are two identical data formats to read from - \code{"psv"}
(flat, pipe-delimited files) and \code{"parquet"} (a newer, faster, columnar
format). The latter is typically faster, but requires the \code{arrow} package
as an additional dependency. Note that this only applies when \code{year} is not
\code{NULL}; all \code{by-site} files are \code{psv} files.}

\item{hourly}{Should hourly means be calculated? The default is \code{TRUE}. If
\code{FALSE} then the raw data are returned, which can be sub-hourly.}

\item{extra}{Should additional columns be returned? The default, \code{FALSE},
returns an opinionated selection of elements that'll be of most interest to
most users. \code{TRUE} will return everything available from the GHCNh.}

\item{abbr_names}{Should column names be abbreviated? When \code{TRUE}, the
default, columns like \code{"wind_direction"} are shortened to \code{"wd"}. When
\code{FALSE}, names will match the raw data, albeit in lower case.}

\item{append_codes}{Logical. Should various codes (e.g., measurement coeds,
qualiy codes, etc.) be appended to the output dataframe?}

\item{codes}{When \code{append_codes} is \code{TRUE}, which codes should be appended to
the dataframe? Any combination of \code{"measurement_code"}, \code{"quality_code"},
\code{"report_type"}, \code{"source_code"}, and/or \code{"source_id"}.}

\item{progress}{Show a progress bar when importing many stations/years?
Defaults to \code{TRUE} in interactive R sessions. Passed to \code{.progress} in
\code{\link[purrr:map]{purrr::map()}} and/or \code{\link[purrr:pmap]{purrr::pmap()}}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function flexibly accesses meteorological data from the GHCNh database.
Users can provide any number of years and stations, and fully control the
sorts of data flag codes that are returned with the data. By default, column
names are shortened for easier use in R, but longer, more descriptive names
can be requested.
}
\seealso{
Other GHCN functions: 
\code{\link{import_ghcn_countries}()},
\code{\link{import_ghcn_daily}()},
\code{\link{import_ghcn_inventory}()},
\code{\link{import_ghcn_monthly_temp}()},
\code{\link{import_ghcn_stations}()}
}
\author{
Jack Davison
}
\concept{GHCN functions}
