% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_isd_stations.R
\name{import_isd_stations_live}
\alias{import_isd_stations_live}
\title{Obtain site meta data from NOAA server}
\usage{
import_isd_stations_live(...)
}
\arguments{
\item{...}{Currently unused.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Download all NOAA meta data, allowing for re-use and direct querying.
}
\examples{
\dontrun{
meta <- import_isd_stations_live()
head(meta)
}
}
\seealso{
Other NOAA ISD functions: 
\code{\link{import_isd_hourly}()},
\code{\link{import_isd_lite}()},
\code{\link{import_isd_stations}()}
}
\concept{NOAA ISD functions}
