% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherCodes.R
\docType{data}
\name{weatherCodes}
\alias{weatherCodes}
\title{Codes for weather types}
\description{
This data frame consists of the weather description codes used in the ISD. It is not of general use to most users.
}
\examples{


## basic structure
head(weatherCodes)


}
\keyword{datasets}
