\name{worldmet-package}
\alias{worldmet-package}
\alias{worldmet}
\docType{package}
\title{
\packageTitle{worldmet}
}
\description{
\packageDescription{worldmet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{worldmet}
\packageIndices{worldmet}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{worldmet}

Maintainer: \packageMaintainer{worldmet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
