% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmet-package.R
\docType{package}
\name{worldmet-package}
\alias{worldmet}
\alias{worldmet-package}
\title{worldmet: Import Surface Meteorological Data from NOAA Integrated Surface Database (ISD)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to import data from more than 30,000 surface meteorological sites around the world managed by the National Oceanic and Atmospheric Administration (NOAA) Global Historical Climatology Network.
}
\details{
This package contains functions to import surface meteorological
data from over 30,000 sites around the world. These data are curated by
NOAA as part of the Global Historical Climatology Network (GHCN).

If you access these data using the \code{{worldmet}} package please give full
acknowledgement to NOAA. Users should also take a note of the usage
restrictions.

These data work well with the \code{{openair}} package that has been developed to
analyse air pollution data.
}
\references{
For general information of the GHCN see
\url{https://www.ncei.noaa.gov/products/global-historical-climatology-network-hourly}
}
\seealso{
See \url{https://github.com/davidcarslaw/openair} for information on the
related \code{openair} package.
}
\author{
\strong{Maintainer}: David Carslaw \email{david.carslaw@york.ac.uk}

Authors:
\itemize{
  \item Jack Davison \email{jack.davison@ricardo.com}
}

}
\keyword{internal}
